add: ABC258

+++ ABC258/b/main.cpp
@@ -8,20 +8,20 @@
 using namespace std;
 
 signed main() {
-    ll N;
+    int N;
     cin >> N;
     vector<string> A(N);
     rep(i, N) {
         cin >> A[i];
     }
-    vector<ll> move_hor = {1, 1, 1, 0, 0, -1, -1, -1};
-    vector<ll> move_ver = {1, 0, -1, 1, -1, 1, 0, -1};
-    ll max_number = 0;
+    vector<int> move_hor = {1, 1, 1, 0, 0, -1, -1, -1};
+    vector<int> move_ver = {1, 0, -1, 1, -1, 1, 0, -1};
+    int max_number = 0;
     rep(i, N) {
         rep(j, N) {
             rep(k, 8) {
-                ll keta = 1;
-                ll loaded_number = 0;
+                int keta = 1;
+                int loaded_number = 0;
                 rep(l, N) {
                     loaded_number += (keta * (A[(N + i + move_ver[k] * l) % N][(N + j + move_hor[k] * l) % N] - '0'));
                     keta *= 10;
